Moved Kubernetes adapter to its own package
